swamp-fox-summit-actor/requirements.txt
apify==2.0.1
httpx==0.27.2
aiohttp==3.10.5
beautifulsoup4==4.12.3
tldextract==5.1.2
dnspython==2.6.1
gspread==6.1.4
google-auth==2.34.0
azure-identity==1.17.1
azure-storage-file-datalake==12.16.0
pandas==2.2.2
pyarrow==17.0.0
deltalake==0.20.1
email-validator==2.2.0
phonenumbers==8.13.44
